+++ internal/models/users.go
@@ -1,8 +1,8 @@
 package models
 
 type Users struct {
-	ID        int    `json:"id" db:"id"`
+	ID        uint64 `json:"id" db:"id"`
 	PublicKey string `json:"public_key" db:"public_key"`
 	Name      string `json:"name" db:"name"`
-	JoinDate  int    `json:"join_date" db:"join_date"`
+	JoinDate  uint32 `json:"join_date" db:"join_date"`
 }
